Flytt visningslogikk inn til situasjon-komponenten
